use int16 for read qual scores because int8 can overflow in some cases

+++ haplokit/io/bam.py
@@ -108,7 +108,7 @@ def extract_read_variants(
                     # default is array of gaps with qual of 0
                     chars = np.empty(n_positions, dtype='U1')
                     chars[:] = '-'
-                    quals = np.zeros(n_positions, dtype=np.int8)
+                    quals = np.zeros(n_positions, dtype=np.int16)
                     sample_data[read.qname] = [chars, quals]
                     
                 else:
@@ -164,7 +164,7 @@ def add_nan_read_if_empty(locus, symbols, quals):
     assert np.size(symbols) == np.size(quals)
     if np.size(symbols) == 0:
         symbols = np.array([['-'] * len(locus.variants) ])
-        quals = np.zeros((1, len(locus.variants)), dtype=np.int8)
+        quals = np.zeros((1, len(locus.variants)), dtype=np.int16)
     return symbols, quals
